+++ code-types/ExtractAst.js
@@ -9,7 +9,7 @@ var ExtractAst;
      * @param childTypes the list of 'allTypeDefs' names to extract from
      * @param allTypeDefs a map of all type definition names and CodeAst classes
      * @param transformTypeName an optional transformer for the returned type name map
-     * @return a map of type names used
+     * @return a map of types found
      */
     function extractInheritedTypeNames(childTypes, allTypeDefs, transformTypeName) {
         var typesUsed = {};
@@ -23,7 +23,8 @@ var ExtractAst;
                     for (var k = 0, sizeK = extendTypes.length; k < sizeK; k++) {
                         var t1 = transformTypeName != null ? transformTypeName(extendTypes[k]) : extendTypes[k];
                         if (t1 != null) {
-                            typesUsed[t1] = true;
+                            var tu1 = typesUsed[t1] || (typesUsed[t1] = []);
+                            tu1.push({ class: classDef, extendType: extendTypes[k] });
                         }
                     }
                 }
@@ -33,7 +34,8 @@ var ExtractAst;
                         for (var p = 0, sizeP = implementTypes.length; p < sizeP; p++) {
                             var t2 = transformTypeName != null ? transformTypeName(implementTypes[p]) : implementTypes[p];
                             if (t2 != null) {
-                                typesUsed[t2] = true;
+                                var tu2 = typesUsed[t2] || (typesUsed[t2] = []);
+                                tu2.push({ class: classDef, implementType: implementTypes[p] });
                             }
                         }
                     }
@@ -47,6 +49,7 @@ var ExtractAst;
      * @param childTypes
      * @param typeDefs
      * @param includePrimitiveTypes
+     * @return a map of types found
      */
     function extractFieldTypeNames(childTypes, typeDefs, includePrimitiveTypes, transformTypeName) {
         var typesUsed = {};
@@ -62,7 +65,8 @@ var ExtractAst;
                         var typeName = transformTypeName != null ? transformTypeName(fieldTypes[m]) : fieldTypes[m]; // returns null for types not starting with upper-case letter (which automatically excludes most primitives)
                         if (typeName != null) {
                             if (includePrimitiveTypes || (!TypeConverter.isPrimitive(typeName) && !TypeConverter.isCore(typeName))) {
-                                typesUsed[typeName] = true;
+                                var tu1 = typesUsed[typeName] || (typesUsed[typeName] = []);
+                                tu1.push({ class: classDef, field: fields[k] });
                             }
                         }
                     }