+++ code-types/TypeConverter.ts
@@ -102,13 +102,20 @@ module TypeConverter {
             case "short":
             case "ushort":
             case "int":
+            case "int16":
+            case "int32":
+            case "int64":
             case "uint":
+            case "uint16":
+            case "uint32":
+            case "uint64":
             case "long":
             case "ulong":
             case "float":
             case "double":
             case "decimal":
             case "real":
+            case "single":
             case "number":
                 return true;
             default:
@@ -126,6 +133,7 @@ module TypeConverter {
         switch (typeName) {
             case "date":
             case "datetime":
+            case "object":
             case "string":
                 return true;
             default: